+++ example/android/app/src/main/kotlin/flutter/persian/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package flutter.persian.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}