+++ lib/src/pdate_picker_dialog.dart
@@ -4,9 +4,7 @@
 
 import 'dart:math' as math;
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:persian_datetime_picker/src/date/shamsi_date.dart';
 
 import 'pcalendar_date_picker.dart';
@@ -39,9 +37,9 @@ borderRadius: BorderRadius.all(Radius.circular(4.0)));
 /// be non-null.
 ///
 /// An optional [initialEntryMode] argument can be used to display the date
-/// picker in the [DatePickerEntryMode.calendar] (a calendar month grid)
-/// or [DatePickerEntryMode.input] (a text input field) mode.
-/// It defaults to [DatePickerEntryMode.calendar] and must be non-null.
+/// picker in the [PDatePickerEntryMode.calendar] (a calendar month grid)
+/// or [PDatePickerEntryMode.input] (a text input field) mode.
+/// It defaults to [PDatePickerEntryMode.calendar] and must be non-null.
 ///
 /// An optional [selectableDayPredicate] function can be passed in to only allow
 /// certain days for selection. If provided, only the days that
@@ -86,7 +84,7 @@ Future<Jalali?> showPersianDatePicker({
   required Jalali initialDate,
   required Jalali firstDate,
   required Jalali lastDate,
-  DatePickerEntryMode initialEntryMode = DatePickerEntryMode.calendar,
+  PDatePickerEntryMode initialEntryMode = PDatePickerEntryMode.calendar,
   PSelectableDayPredicate? selectableDayPredicate,
   ShapeBorder selectedShape = const RoundedRectangleBorder(
       borderRadius: BorderRadius.all(Radius.circular(4.0))),
@@ -101,8 +99,8 @@ Future<Jalali?> showPersianDatePicker({
   PDatePickerMode initialDatePickerMode = PDatePickerMode.day,
   String? errorFormatText,
   String? errorInvalidText,
-  String fieldHintText = "##/##/####",
-  String fieldLabelText = "ورود تاریخ",
+  String fieldHintText = '##/##/####',
+  String fieldLabelText = 'ورود تاریخ',
 }) async {
   shapeBorder = selectedShape;
   assert(context != null);
@@ -120,9 +118,6 @@ Future<Jalali?> showPersianDatePicker({
       'initialDate $initialDate must be on or before lastDate $lastDate.');
   assert(selectableDayPredicate == null || selectableDayPredicate(initialDate),
       'Provided initialDate $initialDate must satisfy provided selectableDayPredicate.');
-  assert(initialEntryMode != null);
-  assert(useRootNavigator != null);
-  assert(initialDatePickerMode != null);
   assert(debugCheckHasMaterialLocalizations(context));
 
   Widget dialog = _DatePickerDialog(
@@ -172,7 +167,7 @@ class _DatePickerDialog extends StatefulWidget {
     required Jalali initialDate,
     required Jalali firstDate,
     required Jalali lastDate,
-    this.initialEntryMode = DatePickerEntryMode.calendar,
+    this.initialEntryMode = PDatePickerEntryMode.calendar,
     this.selectableDayPredicate,
     this.cancelText,
     this.confirmText,
@@ -182,14 +177,9 @@ class _DatePickerDialog extends StatefulWidget {
     this.errorInvalidText,
     this.fieldHintText,
     this.fieldLabelText,
-  })  : assert(initialDate != null),
-        assert(firstDate != null),
-        assert(lastDate != null),
-        initialDate = utils.dateOnly(initialDate),
+  })  : initialDate = utils.dateOnly(initialDate),
         firstDate = utils.dateOnly(firstDate),
         lastDate = utils.dateOnly(lastDate),
-        assert(initialEntryMode != null),
-        assert(initialCalendarMode != null),
         super(key: key) {
     assert(!this.lastDate.isBefore(this.firstDate),
         'lastDate ${this.lastDate} must be on or after firstDate ${this.firstDate}.');
@@ -212,7 +202,7 @@ class _DatePickerDialog extends StatefulWidget {
   /// The latest allowable [Jalali] that the user can select.
   final Jalali lastDate;
 
-  final DatePickerEntryMode initialEntryMode;
+  final PDatePickerEntryMode initialEntryMode;
 
   /// Function to provide full control over which [Jalali] can be selected.
   final PSelectableDayPredicate? selectableDayPredicate;
@@ -244,7 +234,7 @@ class _DatePickerDialog extends StatefulWidget {
 }
 
 class _DatePickerDialogState extends State<_DatePickerDialog> {
-  DatePickerEntryMode? _entryMode;
+  PDatePickerEntryMode? _entryMode;
   Jalali? _selectedDate;
   late bool _autoValidate;
   final GlobalKey _calendarPickerKey = GlobalKey();
@@ -259,7 +249,7 @@ class _DatePickerDialogState extends State<_DatePickerDialog> {
   }
 
   void _handleOk() {
-    if (_entryMode == DatePickerEntryMode.input) {
+    if (_entryMode == PDatePickerEntryMode.input ||  _entryMode == DatePickerEntryMode.inputOnly) {
       final FormState form = _formKey.currentState!;
       if (!form.validate()) {
         setState(() => _autoValidate = true);
@@ -277,15 +267,17 @@ class _DatePickerDialogState extends State<_DatePickerDialog> {
   void _handelEntryModeToggle() {
     setState(() {
       switch (_entryMode) {
-        case DatePickerEntryMode.calendar:
+        case PDatePickerEntryMode.calendar:
           _autoValidate = false;
-          _entryMode = DatePickerEntryMode.input;
+          _entryMode = PDatePickerEntryMode.input;
           break;
-        case DatePickerEntryMode.input:
+        case PDatePickerEntryMode.input:
           _formKey.currentState!.save();
-          _entryMode = DatePickerEntryMode.calendar;
+          _entryMode = PDatePickerEntryMode.calendar;
           break;
-        default:
+        case PDatePickerEntryMode.calendarOnly:
+        case PDatePickerEntryMode.inputOnly:
+          assert(false, 'Can not change entry mode from _entryMode');
           break;
       }
     });
@@ -298,23 +290,22 @@ class _DatePickerDialogState extends State<_DatePickerDialog> {
   Size? _dialogSize(BuildContext context) {
     final Orientation orientation = MediaQuery.of(context).orientation;
     switch (_entryMode) {
-      case DatePickerEntryMode.calendar:
+      case PDatePickerEntryMode.calendar:
+      case PDatePickerEntryMode.calendarOnly:
         switch (orientation) {
           case Orientation.portrait:
             return _calendarPortraitDialogSize;
           case Orientation.landscape:
             return _calendarLandscapeDialogSize;
         }
-        break;
-      case DatePickerEntryMode.input:
+      case PDatePickerEntryMode.input:
+      case PDatePickerEntryMode.inputOnly:
         switch (orientation) {
           case Orientation.portrait:
             return _inputPortraitDialogSize;
           case Orientation.landscape:
             return _inputLandscapeDialogSize;
         }
-        break;
-
       default:
         return null;
     }
@@ -324,8 +315,6 @@ class _DatePickerDialogState extends State<_DatePickerDialog> {
   Widget build(BuildContext context) {
     final ThemeData theme = Theme.of(context);
     final ColorScheme colorScheme = theme.colorScheme;
-    final MaterialLocalizations localizations =
-        MaterialLocalizations.of(context);
     final Orientation orientation = MediaQuery.of(context).orientation;
     final TextTheme textTheme = theme.textTheme;
     // Constrain the textScaleFactor to the largest supported value to prevent
@@ -341,64 +330,92 @@ class _DatePickerDialogState extends State<_DatePickerDialog> {
         ? colorScheme.onPrimary
         : colorScheme.onSurface;
     final TextStyle? dateStyle = orientation == Orientation.landscape
-        ? textTheme.headline5?.copyWith(color: dateColor)
-        : textTheme.headline4?.copyWith(color: dateColor);
+        ? textTheme.subtitle1?.copyWith(color: dateColor)
+        : textTheme.headline5?.copyWith(color: dateColor);
 
     final Widget actions = ButtonBar(
       buttonTextTheme: ButtonTextTheme.primary,
       layoutBehavior: ButtonBarLayoutBehavior.constrained,
       children: <Widget>[
-        FlatButton(
-          child: Text(widget.cancelText ?? "لغو"),
+        TextButton(
           onPressed: _handleCancel,
+          child: Text(widget.cancelText ?? 'لغو'),
         ),
-        FlatButton(
-          child: Text(widget.confirmText ?? "تایید"),
+        TextButton(
           onPressed: _handleOk,
+          child: Text(widget.confirmText ?? 'تایید'),
         ),
       ],
     );
+
+    PCalendarDatePicker pCalendarDatePicker () {
+      return PCalendarDatePicker(
+        key: _calendarPickerKey,
+        initialDate: _selectedDate!,
+        firstDate: widget.firstDate,
+        lastDate: widget.lastDate,
+        onDateChanged: _handleDateChanged,
+        selectableDayPredicate: widget.selectableDayPredicate,
+        initialCalendarMode: widget.initialCalendarMode,
+      );
+    }
+
+    PInputDatePickerFormField pInputDatePickerFormField() {
+      return PInputDatePickerFormField(
+        initialDate: _selectedDate,
+        firstDate: widget.firstDate,
+        lastDate: widget.lastDate,
+        onDateSubmitted: _handleDateChanged,
+        onDateSaved: _handleDateChanged,
+        selectableDayPredicate: widget.selectableDayPredicate,
+        errorFormatText: widget.errorFormatText,
+        errorInvalidText: widget.errorInvalidText,
+        fieldHintText: widget.fieldHintText,
+        fieldLabelText: widget.fieldLabelText,
+        autofocus: true,
+      );
+    }
 
     late Widget picker;
     IconData? entryModeIcon;
     String? entryModeTooltip;
     switch (_entryMode) {
-      case DatePickerEntryMode.calendar:
-        picker = PCalendarDatePicker(
-          key: _calendarPickerKey,
-          initialDate: _selectedDate!,
-          firstDate: widget.firstDate,
-          lastDate: widget.lastDate,
-          onDateChanged: _handleDateChanged,
-          selectableDayPredicate: widget.selectableDayPredicate,
-          initialCalendarMode: widget.initialCalendarMode,
-        );
+      case PDatePickerEntryMode.calendar:
+        picker = pCalendarDatePicker();
         entryModeIcon = Icons.edit;
         // TODO(darrenaustin): localize 'Switch to input'
         entryModeTooltip = 'Switch to input';
         break;
 
-      case DatePickerEntryMode.input:
+      case PDatePickerEntryMode.calendarOnly:
+        picker = pCalendarDatePicker();
+        // TODO(darrenaustin): localize 'Switch to calendarOnly'
+        entryModeIcon = null;
+        entryModeTooltip = null;
+        break;
+
+      case PDatePickerEntryMode.input:
         picker = Form(
           key: _formKey,
-          autovalidate: _autoValidate,
-          child: PInputDatePickerFormField(
-            initialDate: _selectedDate,
-            firstDate: widget.firstDate,
-            lastDate: widget.lastDate,
-            onDateSubmitted: _handleDateChanged,
-            onDateSaved: _handleDateChanged,
-            selectableDayPredicate: widget.selectableDayPredicate,
-            errorFormatText: widget.errorFormatText,
-            errorInvalidText: widget.errorInvalidText,
-            fieldHintText: widget.fieldHintText,
-            fieldLabelText: widget.fieldLabelText,
-            autofocus: true,
-          ),
+          autovalidateMode: _autoValidate
+              ? AutovalidateMode.onUserInteraction
+              : AutovalidateMode.disabled,
+          child: pInputDatePickerFormField(),
         );
         entryModeIcon = Icons.calendar_today;
         // TODO(darrenaustin): localize 'Switch to calendar'
         entryModeTooltip = 'Switch to calendar';
+        break;
+
+      case PDatePickerEntryMode.inputOnly:
+        picker = Form(
+          key: _formKey,
+          autovalidateMode: _autoValidate
+              ? AutovalidateMode.onUserInteraction
+              : AutovalidateMode.disabled,
+          child: pInputDatePickerFormField(),
+        );
+        // TODO(darrenaustin): localize 'Switch to calendar'
         break;
       default:
         break;
@@ -419,6 +436,15 @@ class _DatePickerDialogState extends State<_DatePickerDialog> {
     final Size dialogSize = _dialogSize(context)! * textScaleFactor;
     final DialogTheme dialogTheme = Theme.of(context).dialogTheme;
     return Dialog(
+      insetPadding:
+          const EdgeInsets.symmetric(horizontal: 16.0, vertical: 24.0),
+      // The default dialog shape is radius 2 rounded rect, but the spec has
+      // been updated to 4, so we will use that here for the Date Picker, but
+      // only if there isn't one provided in the theme.
+      shape: dialogTheme.shape ??
+          const RoundedRectangleBorder(
+              borderRadius: BorderRadius.all(Radius.circular(4.0))),
+      clipBehavior: Clip.antiAlias,
       child: Directionality(
         textDirection: TextDirection.rtl,
         child: AnimatedContainer(