+++ lib/src/ptime_picker.dart
@@ -4,7 +4,6 @@ import 'dart:math' as math;
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 
 // Examples can assume:
 // BuildContext context;
@@ -45,13 +44,15 @@ ShapeBorder selectedShape = _kDefaultShape;
 enum PTimePickerEntryMode {
   /// Tapping/dragging on a clock dial.
   dial,
-
+  /// Tapping/dragging on a clock dial only.
+  dialOnly,
   /// Text input.
   input,
+  /// Text input only.
+  inputOnly,
 }
 
 /// Provides properties for rendering time picker header fragments.
-@immutable
 class _TimePickerFragmentContext {
   const _TimePickerFragmentContext({
     required this.selectedTime,
@@ -59,11 +60,7 @@ class _TimePickerFragmentContext {
     required this.onTimeChange,
     required this.onModeChange,
     required this.use24HourDials,
-  })  : assert(selectedTime != null),
-        assert(mode != null),
-        assert(onTimeChange != null),
-        assert(onModeChange != null),
-        assert(use24HourDials != null);
+  });
 
   final TimeOfDay selectedTime;
   final _TimePickerMode mode;
@@ -81,10 +78,7 @@ class _TimePickerHeader extends StatelessWidget {
     required this.onChanged,
     required this.use24HourDials,
     required this.helpText,
-  })  : assert(selectedTime != null),
-        assert(mode != null),
-        assert(orientation != null),
-        assert(use24HourDials != null);
+  });
 
   final TimeOfDay selectedTime;
   final _TimePickerMode mode;
@@ -127,7 +121,7 @@ class _TimePickerHeader extends StatelessWidget {
         controls = Column(
           children: <Widget>[
             const SizedBox(height: 16.0),
-            Container(
+            SizedBox(
               height: kMinInteractiveDimension * 2,
               child: Row(
                 children: <Widget>[
@@ -186,7 +180,7 @@ class _TimePickerHeader extends StatelessWidget {
                   orientation: orientation,
                   onChanged: onChanged,
                 ),
-              Container(
+              SizedBox(
                 height: kMinInteractiveDimension * 2,
                 child: Row(
                   // Hour/minutes should not change positions in RTL locales.
@@ -238,9 +232,7 @@ class _HourMinuteControl extends StatelessWidget {
     required this.text,
     required this.onTap,
     required this.isSelected,
-  })  : assert(text != null),
-        assert(onTap != null),
-        assert(isSelected != null);
+  });
 
   final String text;
   final GestureTapCallback onTap;
@@ -270,7 +262,7 @@ class _HourMinuteControl extends StatelessWidget {
     final Set<MaterialState> states = isSelected
         ? <MaterialState>{MaterialState.selected}
         : <MaterialState>{};
-    return Container(
+    return SizedBox(
       height: _kTimePickerHeaderControlHeight,
       child: Material(
         color: MaterialStateProperty.resolveAs(backgroundColor, states),
@@ -383,7 +375,6 @@ class _StringFragment extends StatelessWidget {
       case TimeOfDayFormat.frenchCanadian:
         return 'h';
     }
-    return '';
   }
 
   @override
@@ -523,8 +514,6 @@ class _DayPeriodControl extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final MaterialLocalizations materialLocalizations =
-        MaterialLocalizations.of(context);
     final ColorScheme colorScheme = Theme.of(context).colorScheme;
     final TimePickerThemeData timePickerTheme = TimePickerTheme.of(context);
     final bool isDark = colorScheme.brightness == Brightness.dark;
@@ -583,7 +572,7 @@ class _DayPeriodControl extends StatelessWidget {
           selected: amSelected,
           child: Center(
             child: Text(
-              "ق.ظ",
+              'ق.ظ',
               style: amStyle,
               textScaleFactor: buttonTextScaleFactor,
             ),
@@ -600,7 +589,7 @@ class _DayPeriodControl extends StatelessWidget {
           selected: pmSelected,
           child: Center(
             child: Text(
-              "ب.ظ",
+              'ب.ظ',
               style: pmStyle,
               textScaleFactor: buttonTextScaleFactor,
             ),
@@ -616,7 +605,7 @@ class _DayPeriodControl extends StatelessWidget {
         result = _DayPeriodInputPadding(
           minSize: const Size(width, kMinInteractiveDimension * 2),
           orientation: orientation,
-          child: Container(
+          child: SizedBox(
             width: width,
             height: _kTimePickerHeaderControlHeight,
             child: Material(
@@ -643,7 +632,7 @@ class _DayPeriodControl extends StatelessWidget {
         result = _DayPeriodInputPadding(
           minSize: const Size(0.0, kMinInteractiveDimension),
           orientation: orientation,
-          child: Container(
+          child: SizedBox(
             height: 40.0,
             child: Material(
               clipBehavior: Clip.antiAlias,
@@ -826,7 +815,7 @@ class _DialPainter extends CustomPainter {
     required this.theta,
     required this.textDirection,
     required this.selectedValue,
-  }) : super(repaint: PaintingBinding.instance!.systemFonts);
+  }) : super(repaint: PaintingBinding.instance.systemFonts);
 
   final List<_TappableLabel>? primaryLabels;
   final List<_TappableLabel>? secondaryLabels;
@@ -914,9 +903,7 @@ class _Dial extends StatefulWidget {
     required this.use24HourDials,
     required this.onChanged,
     required this.onHourSelected,
-  })  : assert(selectedTime != null),
-        assert(mode != null),
-        assert(use24HourDials != null);
+  });
 
   final TimeOfDay selectedTime;
   final _TimePickerMode mode;
@@ -940,7 +927,7 @@ class _DialState extends State<_Dial> with SingleTickerProviderStateMixin {
     _theta = _thetaController!
         .drive(CurveTween(curve: standardEasing))
         .drive(_thetaTween!)
-          ..addListener(() => setState(() {/* _theta.value has changed */}));
+      ..addListener(() => setState(() {/* _theta.value has changed */}));
   }
 
   late ThemeData themeData;
@@ -1030,7 +1017,6 @@ class _DialState extends State<_Dial> with SingleTickerProviderStateMixin {
   TimeOfDay _notifyOnChangedIfNeeded({bool roundMinutes = false}) {
     final TimeOfDay current =
         _getTimeForTheta(_theta.value, roundMinutes: roundMinutes);
-    if (widget.onChanged == null) return current;
     if (current != widget.selectedTime) widget.onChanged(current);
     return current;
   }
@@ -1076,9 +1062,7 @@ class _DialState extends State<_Dial> with SingleTickerProviderStateMixin {
     _center = null;
     _animateTo(_getThetaForTime(widget.selectedTime));
     if (widget.mode == _TimePickerMode.hour) {
-      if (widget.onHourSelected != null) {
-        widget.onHourSelected();
-      }
+      widget.onHourSelected();
     }
   }
 
@@ -1096,9 +1080,7 @@ class _DialState extends State<_Dial> with SingleTickerProviderStateMixin {
         _announceToAccessibility(
             context, localizations.formatDecimal(newTime.hourOfPeriod));
       }
-      if (widget.onHourSelected != null) {
-        widget.onHourSelected();
-      }
+      widget.onHourSelected();
     } else {
       _announceToAccessibility(
           context, localizations.formatDecimal(newTime.minute));
@@ -1221,7 +1203,7 @@ class _DialState extends State<_Dial> with SingleTickerProviderStateMixin {
       ];
 
   List<_TappableLabel> _buildMinutes(TextTheme textTheme, Color color) {
-    const List<TimeOfDay> _minuteMarkerValues = <TimeOfDay>[
+    const List<TimeOfDay> minuteMarkerValues = <TimeOfDay>[
       TimeOfDay(hour: 0, minute: 0),
       TimeOfDay(hour: 0, minute: 5),
       TimeOfDay(hour: 0, minute: 10),
@@ -1237,7 +1219,7 @@ class _DialState extends State<_Dial> with SingleTickerProviderStateMixin {
     ];
 
     return <_TappableLabel>[
-      for (final TimeOfDay timeOfDay in _minuteMarkerValues)
+      for (final TimeOfDay timeOfDay in minuteMarkerValues)
         _buildTappableLabel(
           textTheme,
           color,
@@ -1274,19 +1256,18 @@ class _DialState extends State<_Dial> with SingleTickerProviderStateMixin {
           selectedDialValue = widget.selectedTime.hour;
           primaryLabels = _build24HourRing(theme.textTheme, primaryLabelColor);
           secondaryLabels =
-              _build24HourRing(theme.accentTextTheme, secondaryLabelColor);
+              _build24HourRing(theme.textTheme, secondaryLabelColor);
         } else {
           selectedDialValue = widget.selectedTime.hourOfPeriod;
           primaryLabels = _build12HourRing(theme.textTheme, primaryLabelColor);
           secondaryLabels =
-              _build12HourRing(theme.accentTextTheme, secondaryLabelColor);
+              _build12HourRing(theme.textTheme, secondaryLabelColor);
         }
         break;
       case _TimePickerMode.minute:
         selectedDialValue = widget.selectedTime.minute;
         primaryLabels = _buildMinutes(theme.textTheme, primaryLabelColor);
-        secondaryLabels =
-            _buildMinutes(theme.accentTextTheme, secondaryLabelColor);
+        secondaryLabels = _buildMinutes(theme.textTheme, secondaryLabelColor);
         break;
     }
 
@@ -1319,9 +1300,7 @@ class _TimePickerInput extends StatefulWidget {
     required this.initialSelectedTime,
     required this.helpText,
     required this.onChanged,
-  })  : assert(initialSelectedTime != null),
-        assert(onChanged != null),
-        super(key: key);
+  }) : super(key: key);
 
   /// The time initially selected when the dialog is shown.
   final TimeOfDay initialSelectedTime;
@@ -1347,10 +1326,6 @@ class _TimePickerInputState extends State<_TimePickerInput> {
   }
 
   int? _parseHour(String value) {
-    if (value == null) {
-      return null;
-    }
-
     int? newHour = int.tryParse(value);
     if (newHour == null) {
       return null;
@@ -1374,10 +1349,6 @@ class _TimePickerInputState extends State<_TimePickerInput> {
   }
 
   int? _parseMinute(String value) {
-    if (value == null) {
-      return null;
-    }
-
     final int? newMinute = int.tryParse(value);
     if (newMinute == null) {
       return null;
@@ -1499,7 +1470,7 @@ class _TimePickerInputState extends State<_TimePickerInput> {
                           if (!hourHasError && !minuteHasError)
                             ExcludeSemantics(
                               child: Text(
-                                "ساعت",
+                                'ساعت',
                                 style: theme.textTheme.caption,
                                 maxLines: 1,
                                 overflow: TextOverflow.ellipsis,
@@ -1528,7 +1499,7 @@ class _TimePickerInputState extends State<_TimePickerInput> {
                           if (!hourHasError && !minuteHasError)
                             ExcludeSemantics(
                               child: Text(
-                                "دقیقه",
+                                'دقیقه',
                                 style: theme.textTheme.caption,
                                 maxLines: 1,
                                 overflow: TextOverflow.ellipsis,
@@ -1554,7 +1525,7 @@ class _TimePickerInputState extends State<_TimePickerInput> {
           ),
           if (hourHasError || minuteHasError)
             Text(
-              "زمان معتبر نیست.",
+              'زمان معتبر نیست.',
               style: theme.textTheme.bodyText2!
                   .copyWith(color: theme.colorScheme.error),
             )
@@ -1768,8 +1739,7 @@ class _TimePickerDialog extends StatefulWidget {
     required this.confirmText,
     required this.helpText,
     this.initialEntryMode = PTimePickerEntryMode.dial,
-  })  : assert(initialTime != null),
-        super(key: key);
+  }) : super(key: key);
 
   /// The time initially selected when the dialog is shown.
   final TimeOfDay initialTime;
@@ -1861,8 +1831,10 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
           _formKey.currentState!.save();
           _entryMode = PTimePickerEntryMode.dial;
           break;
-        default:
-          break;
+        case PTimePickerEntryMode.dialOnly:
+        case PTimePickerEntryMode.inputOnly:
+          FlutterError('Can not change entry mode from $_entryMode');
+        break;
       }
     });
   }
@@ -1920,7 +1892,7 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
   }
 
   void _handleOk() {
-    if (_entryMode == PTimePickerEntryMode.input) {
+    if (_entryMode == PTimePickerEntryMode.input || _entryMode == PTimePickerEntryMode.inputOnly) {
       final FormState form = _formKey.currentState!;
       if (!form.validate()) {
         setState(() {
@@ -1946,6 +1918,7 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
     late double timePickerHeight;
     switch (_entryMode) {
       case PTimePickerEntryMode.dial:
+      case PTimePickerEntryMode.dialOnly:
         switch (orientation) {
           case Orientation.portrait:
             timePickerWidth = _kTimePickerWidthPortrait;
@@ -1962,13 +1935,11 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
                     : _kTimePickerHeightLandscapeCollapsed;
             break;
         }
-
         break;
       case PTimePickerEntryMode.input:
+      case PTimePickerEntryMode.inputOnly:
         timePickerWidth = _kTimePickerWidthPortrait;
         timePickerHeight = _kTimePickerHeightInput;
-        break;
-      default:
         break;
     }
     return Size(timePickerWidth, timePickerHeight * textScaleFactor);
@@ -1995,7 +1966,8 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
                 theme.colorScheme.brightness == Brightness.dark ? 1.0 : 0.6,
               ),
           onPressed: _handleEntryModeToggle,
-          icon: Icon(_entryMode == PTimePickerEntryMode.dial
+          icon: _entryMode == PTimePickerEntryMode.dialOnly || _entryMode == PTimePickerEntryMode.dialOnly ? Icon(null) :
+              Icon(_entryMode == PTimePickerEntryMode.dial
               ? Icons.keyboard
               : Icons.access_time),
           tooltip: _entryMode == PTimePickerEntryMode.dial
@@ -2029,6 +2001,7 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
     Widget? picker;
     switch (_entryMode) {
       case PTimePickerEntryMode.dial:
+      case PTimePickerEntryMode.dialOnly:
         final Widget dial = Padding(
           padding: orientation == Orientation.portrait
               ? const EdgeInsets.symmetric(horizontal: 36, vertical: 24)
@@ -2095,9 +2068,12 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
         }
         break;
       case PTimePickerEntryMode.input:
+      case PTimePickerEntryMode.inputOnly:
         picker = Form(
           key: _formKey,
-          autovalidate: _autoValidate,
+          autovalidateMode: _autoValidate
+              ? AutovalidateMode.onUserInteraction
+              : AutovalidateMode.disabled,
           child: SingleChildScrollView(
             child: Column(
               mainAxisSize: MainAxisSize.min,
@@ -2124,7 +2100,7 @@ class _TimePickerDialogState extends State<_TimePickerDialog> {
           theme.colorScheme.surface,
       insetPadding: EdgeInsets.symmetric(
         horizontal: 16.0,
-        vertical: _entryMode == PTimePickerEntryMode.input ? 0.0 : 24.0,
+        vertical: _entryMode == PTimePickerEntryMode.input || _entryMode == PTimePickerEntryMode.inputOnly ? 0.0 : 24.0,
       ),
       child: AnimatedContainer(
         width: dialogSize.width,
@@ -2226,15 +2202,11 @@ Future<TimeOfDay?> showPersianTimePicker({
   TransitionBuilder? builder,
   bool useRootNavigator = true,
   PTimePickerEntryMode initialEntryMode = PTimePickerEntryMode.dial,
-  String cancelText = "لغو",
-  String confirmText = "تایید",
-  String helpText = "انتخاب زمان",
+  String cancelText = 'لغو',
+  String confirmText = 'تایید',
+  String helpText = 'انتخاب زمان',
   RouteSettings? routeSettings,
 }) async {
-  assert(context != null);
-  assert(initialTime != null);
-  assert(useRootNavigator != null);
-  assert(initialEntryMode != null);
   assert(debugCheckHasMaterialLocalizations(context));
   selectedBorderRadius = borderRadius;
   selectedShape = shapeBorder;